Issue: Null scalar values are not valid
Description: In our usage, we encountered cases where null float values, null booleans were being converted to scalar 0 and false values respectively(strings seem fine and return null). Null seems an acceptable value for the client to handle or check of errors if they choose to. I made a few changes in the test cases to show which types needed this change.