More realistic GraphQL vs REST explanation
Hey there! I'm making this PR with the best of intentions, but I'm sure it'll be taken as aggressive. Not the goal. 

Recently I've been [trying to explain to people](https://philsturgeon.uk/api/2017/01/24/graphql-vs-rest-overview/) the differences between GraphQL and REST and debunk some of the "GraphQL is magically better" stuff going around. Your introduction is furthering the issue. 

Generally, GraphQL and REST are just different, and pretending that one is better is going to confuse people.

> self-checks embedded on the ground level of your backend architecture

What does this mean

> reusable API for different client versions and devices, i.e. no more need in maintaining "/v1" and "/v2"

This was never recommended in REST, in fact the opposite was recommended. [Evolvability is king for both](https://www.mnot.net/blog/2012/12/04/api-evolution).

> easily generated documentation and incredibly intuitive way to explore created API

Yeah! Love it, but their schemas are based on concepts like JSON Schema, JSON LD, OpenAPI, etc, and exploring the API is what HATEOAS provides, which is a REST concept. 

> once your architecture is complete – most client-based changes does not require backend modifications

👍

I had a go at making it a bit more accurate for you, I hope you like it!